+++ src/openweather_sdk/__init__.py
@@ -13,6 +13,7 @@ from openweather_sdk.exceptions import (
 )
 from openweather_sdk.globals import _LANGUAGES, _SPECIFIC_CACHES, _UNITS, _WORK_MODES
 from openweather_sdk.json_processor import _JSONProcessor
+from openweather_sdk.rest.airpollution import _AirPollutionAPI
 from openweather_sdk.rest.forecast import _ForecastAPI
 from openweather_sdk.rest.geocoding import _GeocodingAPI
 from openweather_sdk.rest.openweather import _OpenWeather
@@ -310,11 +311,36 @@ class Client:
 
         if self.cache_size:
             with self.lock:
-                self.cache["forecast_16_days"]._add_info(lon, lat, forecast)
+                self.cache["forecast_30_days"]._add_info(lon, lat, forecast)
                 logger.info(
                     f"The client {self} has received data about 30 days forecast: {forecast}"
                 )
                 return forecast
+
+    def _get_location_current_air_pollution(self, location):
+        lon, lat = self._get_location_coordinates(location)
+        return self._get_current_air_pollution(lon, lat)
+
+    def _get_zip_code_current_air_pollution(self, zip_code):
+        lon, lat = self._get_zip_code_coordinates(zip_code)
+        return self._get_current_air_pollution(lon, lat)
+
+    def _get_current_air_pollution(self, lon, lat):
+        if self.cache_size:
+            with self.lock:
+                if self.cache["current_air_pollution"]._is_relevant_info(lon, lat):
+                    return self.cache["current_air_pollution"]._get_info(lon, lat)
+
+        air_pollution_api = _AirPollutionAPI(lon=lon, lat=lat, appid=self.token)
+        air_pollution = air_pollution_api._get_current_air_pollution()
+
+        if self.cache_size:
+            with self.lock:
+                self.cache["current_air_pollution"]._add_info(lon, lat, air_pollution)
+                logger.info(
+                    f"The client {self} has received data about current air pollution: {air_pollution}"
+                )
+                return air_pollution
 
     def get_location_weather(self, location, compact_mode=True):
         """
@@ -549,3 +575,37 @@ class Client:
                     "You need to specify zip code as a string"
                 )
             return self._get_zip_code_forecast_daily_30_days(zip_code)
+
+    def current_air_pollution(self, location=None, zip_code=None):
+        """
+        Returns current air pollution in a specified location.
+        The location can be provided either as a combination of city name,
+        state code (for the US), and country code separated by commas, or
+        as a combination of zip/post code and country code separated by commas.
+        Please ensure the usage of ISO 3166 country codes.
+
+        Args:
+            location (str, optional): city name, state code (only for the US) and country code divided by comma.
+            zip_code (str, optional): zip/post code and country code divided by comma.
+        """
+        logger.info(
+            f"The client {self} is being requested the current air pollution in the location {location or zip_code}..."
+        )
+        if not self.is_alive:
+            raise ClientDoesntExistException(self)
+        if not location and not zip_code:
+            raise InvalidLocationException(
+                "You need to specify the location or postal code."
+            )
+        if location:
+            if not isinstance(location, str):
+                raise InvalidLocationException(
+                    "You need to specify the location as a string."
+                )
+            return self._get_location_current_air_pollution(location)
+        if zip_code:
+            if not isinstance(zip_code, str):
+                raise InvalidLocationException(
+                    "You need to specify zip code as a string"
+                )
+            return self._get_zip_code_current_air_pollution(zip_code)